move stub client to client_test.go and make private
